+++ NEWS.md
@@ -1,3 +1,4 @@
 # Racmacs 1.1.10
 
 * Added a `NEWS.md` file to track changes to the package.
+* Add test and correction for error where coordinates of underconstrained points were set to NaN for the first optimization run only, not all.